# Avoid `value_for_database` if attribute not updated